+++ src/program/mod.rs
@@ -2,10 +2,12 @@ use std::fmt;
 use std::error::Error;
 use std::sync::Mutex;
 
+pub use self::compute::ComputeShader;
 pub use self::program::Program;
 pub use self::reflection::{Uniform, UniformBlock, UniformBlockMember, OutputPrimitives};
 pub use self::reflection::{Attribute, TransformFeedbackVarying, TransformFeedbackBuffer, TransformFeedbackMode};
 
+mod compute;
 mod program;
 mod raw;
 mod reflection;
